requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.0.10
django-cors-headers==4.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
openpyxl==3.1.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7